Fix a constant pool reading bug
Handle > Short.MAX_VALUE constant pool entries by storing unsigned
shorts as ints instead of signed shorts.

I noticed this testing with the full implementation jar for a very large
proto; apparently we've never hit this limit for ijar/hjars.

MOE_MIGRATED_REVID=164506910

+++ java/com/google/turbine/bytecode/AnnotationWriter.java
@@ -112,7 +112,7 @@ public class AnnotationWriter {
     }
   }
 
-  private void writeConst(char tag, short index) {
+  private void writeConst(char tag, int index) {
     output.writeByte(tag);
     output.writeShort(index);
   }
